pyproject.toml
[project]
name = "mcp-sql-server"
version = "0.1.0"
description = "MCP server for SQL Server database interaction"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.2.0",
    "pyodbc>=5.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
mcp-sql-server = "mcp_sql_server.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
strict_equality = true
strict_concatenate = true
check_untyped_defs = true
disallow_untyped_decorators = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_reexport = true
warn_no_return = true

[[tool.mypy.overrides]]
module = [
    "pyodbc.*",
    "mcp.*",
    "dotenv.*",
]
ignore_missing_imports = true
